+++ R/model_info.R
@@ -171,7 +171,19 @@ model_info.logitr <- model_info.mclogit
 
 #' @export
 model_info.maxLik <- function(x, verbose = TRUE, ...) {
-  .make_family(x, verbose = verbose, ...)
+  fitfam <- .safe(eval(get_call(x)$family))
+  if (is.null(fitfam)) {
+    .make_family(x, verbose = verbose, ...)
+  } else {
+    .make_family(
+      x,
+      fitfam = fitfam$family,
+      logit.link = fitfam$link == "logit",
+      link.fun = fitfam$link,
+      verbose = verbose,
+      ...
+    )
+  }
 }
 
 #' @export
@@ -1192,6 +1204,20 @@ model_info.polr <- function(x, ...) {
 
 
 #' @export
+model_info.nestedLogit <- function(x, ...) {
+  faminfo <- stats::binomial(link = "logit")
+  .make_family(
+    x = x,
+    fitfam = faminfo$family,
+    logit.link = TRUE,
+    link.fun = faminfo$link,
+    ...
+  )
+}
+
+
+
+#' @export
 model_info.hglm <- function(x, ...) {
   faminfo <- .safe({
     mc <- get_call(x)$family
@@ -1248,9 +1274,17 @@ model_info.svyolr <- function(x, ...) {
 #' @export
 model_info.gamlss <- function(x, ...) {
   faminfo <- get(x$family[1], asNamespace("gamlss"))()
+  # for ZIBNB family, we have only one value, so next line returns NA
+  fitfam <- faminfo$family[2]
+  if (is.na(fitfam)) {
+    fitfam <- faminfo$family
+    if (is.null(fitfam)) {
+      fitfam <- "unknown"
+    }
+  }
   .make_family(
     x = x,
-    fitfam = faminfo$family[2],
+    fitfam = fitfam,
     logit.link = faminfo$mu.link == "logit",
     link.fun = faminfo$mu.link,
     ...