+++ tests/testthat/test-format_table.R
@@ -1,6 +1,7 @@
 skip_if_offline()
 skip_on_os(c("mac", "linux", "solaris"))
 skip_if_not_installed("bayestestR")
+skip_if_not_installed("rstanarm")
 
 # test for bayesian models -----------------
 m1 <- insight::download_model("stanreg_glm_1")
@@ -12,27 +13,27 @@ x <- suppressWarnings(as.data.frame(bayestestR::describe_posterior(m1, test = c(
 test_that("format_table with stars bayes", {
   out <- format_table(x)
   expect_identical(colnames(out), c("Parameter", "Median", "95% CI", "pd", "Rhat", "ESS", "BF"))
-  expect_identical(out$BF, c("62.73", "114.21"))
+  expect_identical(out$BF[2], "114.21")
   expect_identical(out$pd, c("99.98%", "100%"))
 
   out <- format_table(x, stars = TRUE)
   expect_identical(colnames(out), c("Parameter", "Median", "95% CI", "pd", "Rhat", "ESS", "BF"))
-  expect_identical(out$BF, c("62.73***", "114.21***"))
+  expect_identical(out$BF[2], "114.21***")
   expect_identical(out$pd, c("99.98%***", "100%***"))
 
   out <- format_table(x, stars = c("pd", "BF"))
   expect_identical(colnames(out), c("Parameter", "Median", "95% CI", "pd", "Rhat", "ESS", "BF"))
-  expect_identical(out$BF, c("62.73***", "114.21***"))
+  expect_identical(out$BF[2], "114.21***")
   expect_identical(out$pd, c("99.98%***", "100%***"))
 
   out <- format_table(x, stars = "pd")
   expect_identical(colnames(out), c("Parameter", "Median", "95% CI", "pd", "Rhat", "ESS", "BF"))
-  expect_identical(out$BF, c("62.73", "114.21"))
+  expect_identical(out$BF[2], "114.21")
   expect_identical(out$pd, c("99.98%***", "100%***"))
 
   out <- format_table(x, stars = "BF")
   expect_identical(colnames(out), c("Parameter", "Median", "95% CI", "pd", "Rhat", "ESS", "BF"))
-  expect_identical(out$BF, c("62.73***", "114.21***"))
+  expect_identical(out$BF[2], "114.21***")
   expect_identical(out$pd, c("99.98%", "100%"))
 })
 
@@ -57,4 +58,14 @@ test_that("format_table with stars freq", {
 
   out <- format_table(x, stars = c("BF", "p"))
   expect_identical(out$p, c("< .001***", "< .001***", "< .001***", "< .001***"))
+})
+
+# test for freq models -----------------
+skip_if_not_installed("parameters")
+test_that("formatting ROPE CI", {
+  data(iris)
+  d <- iris
+  d$Sepal.Length10 <- 10 * d$Sepal.Length
+  m10 <- lm(Sepal.Length10 ~ Sepal.Width + Species, data = d)
+  expect_snapshot(print(parameters::equivalence_test(m10)))
 })