+++ tests/testthat/test-get_data.R
@@ -1,5 +1,68 @@
 skip_on_os("mac")
 
+test_that("retrieve from same environment", {
+  foo <- data.frame(x = 1:10, y = 2:11)
+  fit <- lm(y ~ x, data = foo)
+
+  expect_no_warning({
+    cols <- names(get_data(fit))
+  })
+  expect_setequal(cols, c("x", "y"))
+})
+
+test_that("retrieve from correct environment", {
+  foo <- function() {
+    foo <- data.frame(x = 1:10, y = 2:11)
+
+    return(lm(y ~ x, data = foo))
+  }
+
+  # There should be no warning about "Could not recover model data from
+  # environment"
+  expect_no_warning({
+    cols <- names(get_data(foo()))
+  })
+  expect_setequal(cols, c("x", "y"))
+})
+
+test_that("fetch from local, not global, environment", {
+  # See #760. If the local environment has a modified version of data also in
+  # the global environment, we should find the local version first, not the
+  # global version.
+
+  foo <- function() {
+    mtcars$cylinders <- factor(mtcars$cyl)
+
+    return(lm(mpg ~ cylinders + disp, data = mtcars))
+  }
+
+  expect_setequal(
+    names(get_data(foo())),
+    c("mpg", "disp", "cylinders")
+  )
+})
+
+test_that("retrieve from call formula environment", {
+  skip_if_not_installed("AER")
+
+  foo <- function() {
+    d <- data.frame(
+      y = rnorm(100),
+      x = rnorm(100)
+    )
+
+    # find_formula(fit)$conditional happens to not have an environment for tobit
+    # models, so get_data() should check environment(get_call(fit)$formula). See
+    # #666
+    return(AER::tobit(y ~ x, data = d, right = 1.5))
+  }
+
+  expect_setequal(
+    names(get_data(foo())),
+    c("x", "y")
+  )
+})
+
 test_that("lme", {
   skip_if_not_installed("nlme")
   data("Orthodont", package = "nlme")
@@ -28,7 +91,7 @@ test_that("lme4", {
 test_that("additional_variables = TRUE", {
   k <- mtcars
   k$qsec[1:10] <- NA
-  k <<- k
+  k <- k
   mod <- lm(mpg ~ hp, k)
   n1 <- nrow(k)
   n2 <- nrow(insight::get_data(mod))
@@ -42,8 +105,8 @@ test_that("lm", {
   set.seed(1023)
   x <- rnorm(1000, sd = 4)
   y <- cos(x) + rnorm(1000)
-  # fails if we assign this locally
-  dat <<- data.frame(x, y)
+
+  dat <- data.frame(x, y)
   mod1 <- lm(y ~ x, data = dat)
   mod2 <- lm(y ~ cos(x), data = dat)
   expect_equal(get_data(mod1), get_data(mod2), ignore_attr = TRUE)
@@ -75,6 +138,34 @@ test_that("get_data lavaan", {
   m <- lavaan::sem(model, data = PoliticalDemocracy)
   expect_s3_class(get_data(m, verbose = FALSE), "data.frame")
   expect_equal(head(get_data(m, verbose = FALSE)), head(PoliticalDemocracy), ignore_attr = TRUE, tolerance = 1e-3)
+
+  # works when data not in environment
+  holz_data <- lavaan::HolzingerSwineford1939
+  HS.model <- " visual  =~ x1 + x2 + x3
+                textual =~ x4 + x5 + x6
+                speed   =~ x7 + x8 + x9 "
+  m_holz <- lavaan::lavaan(HS.model,
+    data = holz_data, auto.var = TRUE, auto.fix.first = TRUE,
+    auto.cov.lv.x = TRUE
+  )
+
+  out1 <- get_data(m_holz)
+  expect_named(
+    out1,
+    c(
+      "id", "sex", "ageyr", "agemo", "school", "grade", "x1", "x2",
+      "x3", "x4", "x5", "x6", "x7", "x8", "x9"
+    )
+  )
+  expect_identical(nrow(out1), 301L)
+
+  # rm(holz_data)
+  # out2 <- get_data(m_holz)
+  # expect_named(
+  #   out2,
+  #   c("x1", "x2","x3", "x4", "x5", "x6", "x7", "x8", "x9")
+  # )
+  # expect_identical(nrow(out2), 301L)
 })
 
 
@@ -104,7 +195,7 @@ test_that("get_data include weights, even if ones", {
 
 
 test_that("lm with transformations", {
-  d <<- data.frame(
+  d <- data.frame(
     time = as.factor(c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5)),
     group = c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2),
     sum = c(0, 5, 10, 15, 20, 0, 20, 25, 45, 50, 0, 5, 10, 15, 20, 0, 20, 25, 45, 50, 0, 5, 10, 15, 20, 0, 20, 25, 45, 50)
@@ -117,21 +208,29 @@ test_that("lm with transformations", {
 test_that("lm with poly and NA in response", {
   d <- iris
   d[1:25, "Sepal.Length"] <- NA
-  d2 <<- d
+  d2 <- d
   m <- lm(Sepal.Length ~ Species / poly(Petal.Width, 2), data = d2)
   expect_equal(get_data(m), iris[26:150, c("Sepal.Length", "Species", "Petal.Width")], ignore_attr = TRUE)
 })
 
 
 test_that("mgcv", {
-  ## NOTE check back every now and then and see if tests still work
-  skip("works interactively")
   skip_if_not_installed("mgcv")
+
+  # mgcv::gam() deliberately does not keep its environment, so get_data() has to
+  # fall back to the model frame. See
+  # https://github.com/cran/mgcv/blob/a4e69cf44a49c84a41a42e90c86995a843733968/R/mgcv.r#L2156-L2159
   d <- iris
   d$NewFac <- rep(c(1, 2), length.out = 150)
   model <- mgcv::gam(Sepal.Length ~ s(Petal.Length, by = interaction(Species, NewFac)), data = d)
+
+  # There should be two warnings: One for failing to get the data from the
+  # environment, and one for not recovering interaction() accurately
+  expect_warning(expect_warning({
+    model_data <- get_data(model)
+  }))
   expect_equal(
-    head(insight::get_data(model)),
+    head(model_data),
     head(d[c("Sepal.Length", "Petal.Length", "Species", "NewFac")]),
     ignore_attr = TRUE
   )
@@ -185,7 +284,7 @@ test_that("get_data() log transform", {
   set.seed(123)
   x <- abs(rnorm(100, sd = 5)) + 5
   y <- exp(2 + 0.3 * x + rnorm(100, sd = 0.4))
-  dat <<- data.frame(y, x)
+  dat <- data.frame(y, x)
 
   mod <- lm(log(y) ~ log(x), data = dat)
   expect_equal(
@@ -334,4 +433,54 @@ test_that("get_data colnames", {
   expect_true(all(colnames(out) %in% c("mpg", "hp", "cyl")))
   out <- get_data(m, additional_variables = TRUE)
   expect_true("qsec" %in% colnames(out))
+})
+
+
+test_that("get_data works for fixest inside functions", {
+  skip_if_not_installed("fixest")
+  data(mtcars)
+
+  # fit within function
+  fixest_wrapper1 <- function(data) {
+    data$cylinders <- factor(data$cyl)
+    fit <- fixest::feglm(mpg ~ cylinders * disp + hp, data = data)
+    return(fit)
+  }
+  global_fixest1 <- fixest_wrapper1(data = mtcars)
+  data <- mtcars[, c("mpg", "disp")]
+  expect_named(
+    get_data(global_fixest1),
+    c("mpg", "cylinders", "disp", "hp")
+  )
+
+  # fit within function, subset
+  fixest_wrapper2 <- function(data) {
+    data$cylinders <- factor(data$cyl)
+    fit <- fixest::feglm(mpg ~ cylinders * disp + hp, data = data)
+    return(fit)
+  }
+  data <- mtcars
+  global_fixest2 <- fixest_wrapper2(data = data[1:20, ])
+  expect_identical(nrow(get_data(global_fixest2)), 20L)
+  expect_named(
+    get_data(global_fixest2),
+    c("mpg", "cylinders", "disp", "hp")
+  )
+
+  data(mtcars)
+  d_cyl <- mtcars
+  d_cyl$cylinders <- factor(d_cyl$cyl)
+  global_fixest3 <- fixest::feglm(mpg ~ cylinders * disp + hp, data = d_cyl)
+  expect_named(
+    get_data(global_fixest3),
+    c("mpg", "cylinders", "disp", "hp")
+  )
+
+  # regular example
+  data(iris)
+  res <- fixest::feglm(Sepal.Length ~ Sepal.Width + Petal.Length | Species, iris, "poisson")
+  expect_named(
+    get_data(res),
+    c("Sepal.Length", "Sepal.Width", "Petal.Length", "Species")
+  )
 })