+++ tests/testthat/test-get_loglikelihood.R
@@ -94,10 +94,12 @@ test_that("get_loglikelihood - (g)lmer", {
   expect_equal(as.numeric(ll), as.numeric(ll2))
 
   model <- download_model("lmerMod_1")
+  skip_if(is.null(model))
   expect_equal(get_loglikelihood(model, estimator = "REML"), logLik(model, REML = TRUE), tolerance = 0.01, ignore_attr = TRUE)
   expect_equal(get_loglikelihood(model, estimator = "ML"), logLik(model, REML = FALSE), tolerance = 0.01, ignore_attr = TRUE)
 
   model <- download_model("merMod_1")
+  skip_if(is.null(model))
   expect_equal(get_loglikelihood(model, estimator = "REML"), logLik(model, REML = FALSE), tolerance = 0.01, ignore_attr = TRUE)
   expect_equal(get_loglikelihood(model, estimator = "ML"), logLik(model, REML = FALSE), tolerance = 0.01, ignore_attr = TRUE)
 })