+++ tests/testthat/test-speedlm.R
@@ -2,11 +2,10 @@ skip_if_not_installed("speedglm")
 data(iris)
 data(mtcars)
 
-m1 <- speedlm(Sepal.Length ~ Petal.Width + Species, data = iris)
-m2 <-
-  speedlm(log(mpg) ~ log(hp) + cyl + I(cyl^2) + poly(wt, degree = 2, raw = TRUE),
-    data = mtcars
-  )
+m1 <- speedglm::speedlm(Sepal.Length ~ Petal.Width + Species, data = iris)
+m2 <- speedglm::speedlm(log(mpg) ~ log(hp) + cyl + I(cyl^2) + poly(wt, degree = 2, raw = TRUE),
+  data = mtcars
+)
 
 test_that("model_info", {
   expect_true(model_info(m1)$is_linear)
@@ -34,11 +33,11 @@ test_that("find_response", {
 })
 
 test_that("get_response", {
-  expect_equal(get_response(m1), iris$Sepal.Length)
+  expect_equal(get_response(m1), iris$Sepal.Length, ignore_attr = TRUE)
 })
 
 test_that("get_predictors", {
-  expect_equal(colnames(get_predictors(m1)), c("Petal.Width", "Species"))
+  expect_identical(colnames(get_predictors(m1)), c("Petal.Width", "Species"))
 })
 
 test_that("link_inverse", {
@@ -51,11 +50,11 @@ test_that("linkfun", {
 
 test_that("get_data", {
   expect_equal(nrow(get_data(m1)), 150)
-  expect_equal(
+  expect_identical(
     colnames(get_data(m1)),
     c("Sepal.Length", "Petal.Width", "Species")
   )
-  expect_equal(colnames(get_data(m2)), c("mpg", "hp", "cyl", "wt"))
+  expect_identical(colnames(get_data(m2)), c("mpg", "hp", "cyl", "wt"))
 })
 
 test_that("find_formula", {
@@ -77,18 +76,18 @@ test_that("find_formula", {
 })
 
 test_that("find_variables", {
-  expect_equal(
+  expect_identical(
     find_variables(m1),
     list(
       response = "Sepal.Length",
       conditional = c("Petal.Width", "Species")
     )
   )
-  expect_equal(
+  expect_identical(
     find_variables(m1, flatten = TRUE),
     c("Sepal.Length", "Petal.Width", "Species")
   )
-  expect_equal(
+  expect_identical(
     find_variables(m2, flatten = TRUE),
     c("mpg", "hp", "cyl", "wt")
   )
@@ -100,7 +99,7 @@ test_that("n_obs", {
 })
 
 test_that("find_parameters", {
-  expect_equal(
+  expect_identical(
     find_parameters(m1),
     list(
       conditional = c(
@@ -111,7 +110,7 @@ test_that("find_parameters", {
       )
     )
   )
-  expect_equal(
+  expect_identical(
     find_parameters(m2),
     list(
       conditional = c(
@@ -125,7 +124,7 @@ test_that("find_parameters", {
     )
   )
   expect_equal(nrow(get_parameters(m1)), 4)
-  expect_equal(
+  expect_identical(
     get_parameters(m1)$Parameter,
     c(
       "(Intercept)",
@@ -141,14 +140,14 @@ test_that("is_multivariate", {
 })
 
 test_that("find_terms", {
-  expect_equal(
+  expect_identical(
     find_terms(m1),
     list(
       response = "Sepal.Length",
       conditional = c("Petal.Width", "Species")
     )
   )
-  expect_equal(
+  expect_identical(
     find_terms(m2),
     list(
       response = "log(mpg)",
@@ -163,7 +162,7 @@ test_that("find_terms", {
 })
 
 test_that("find_algorithm", {
-  expect_equal(find_algorithm(m1), list(algorithm = "eigen"))
+  expect_identical(find_algorithm(m1), list(algorithm = "eigen"))
 })
 
 test_that("find_statistic", {