+++ NEWS.md
@@ -1,11 +1,29 @@
+# insight 0.19.3
+
+## Changes to functions
+
+* `standardize_names()` and `standardize_column_order()` now also recognize the
+  `"response.level"` column name.
+
+* `get_data()` for _lavaan_ models is now more stable at retrieving model data
+  when this is not avaible in the environment.
+
+## Bug fixes
+
+* `get_random()` now returns the same observations as `get_data()` and correctly
+  removes missing values from the data before returning it.
+
+* `find_parameters()` for marginal effects ignores the `"s.value"` column (which
+  was added in a recent update).
+
+* Fixed issue in `get_response()` for _brms_ models with `trunc()` function in
+  the response variable.
+
 # insight 0.19.2
 
 ## Breaking changes
 
 * The minimum needed R version has been bumped to `3.6`.
-
-* Tests for package *speedglm* were removed, because *speedglm* was archived on
-  CRAN.
 
 * `download_model()` no longer errors when a model object could not be downloaded,
   but instead returns `NULL`. This prevents test failures, and allows to skip
@@ -18,12 +36,28 @@
 
 ## New supported models
 
-* `phylolm` and `phyloglm` (package *phylolm*).
+* `phylolm` and `phyloglm` (package *phylolm*), `nestedLogit` (package *nestedLogit*).
 
 ## Bug fixes
 
+* Fixed issue in `get_variance()` for *glmmTMB* models with rank deficient
+  coefficients.
+
 * Fixed issues in `get_weights()` for `glm` models without weights and `na.action`
   not set to default in the model call.
+
+* `clean_names()` now also removes the `relevel()` pattern.
+
+* Fixed issue in `model_info()` for models of class `gamlss`.
+
+* Fixed problems preventing `get_data()` from locating data defined in
+  non-global environments.
+
+* Fixed issue in `get_predicted()` for variables of class numeric matrix created
+  by `scale()`, which were correctly handled only when `get_data()` failed to
+  find the data in the appropriate environment.
+
+* Fixed issue in `model_info()` for `gee` models from `binomial` families.
 
 # insight 0.19.1